refactor(academy): extract glossary presentation
