+++ packages/mobile-app/src/features/academy/presentation/index.ts
@@ -1,2 +1,3 @@
 export * from "./academy-hub.presenter";
 export * from "./AcademyArticleRenderer";
+export * from "./AcademyGlossarySection";
